requirements.txt
aiofiles==0.6.0
aiomysql==0.0.21
aiosqlite==0.16.1
click==7.1.2
httptools==0.1.1
install==1.3.4
iso8601==0.1.14
Mako==1.1.4
MarkupSafe==1.1.1
multidict==4.7.6
PyMySQL==0.9.3
PyPika==0.44.1
pytz==2020.5
sanic==0.8.3
Sanic-Mako==0.5.0
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
ujson==4.0.2
uvloop==0.15.2
websockets==5.0.1
Werkzeug==1.0.1
Sanic-wtf==0.5.0
